+++ lib/widgets/button_push.dart
@@ -1,21 +1,40 @@
 import 'package:flutter/material.dart';
-import 'package:online_course/utils/app_colors.dart';
+// import 'package:online_course/utils/app_colors.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:online_course/utils/app_colors.dart';
 
 class ButtonPush extends StatelessWidget {
-  final String text;
+  final text;
+  final Color color;
+  final Color textColor;
   const ButtonPush({
-    super.key, required this.text,
+    super.key,
+    this.text,
+    this.color = AppColors.main_Color,
+    this.textColor = AppColors.white_Color,
   });
 
   @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-          color: AppColors.main_Color,
-          borderRadius: BorderRadius.circular(15.r)),
-      padding: EdgeInsets.symmetric(vertical: 15.h, horizontal: 40.w),
-      child: Text(text),
+  Widget build(
+    BuildContext context,
+  ) {
+    return GestureDetector(
+      onTap: () {
+        
+      },
+      child: Container(
+        width: double.infinity,
+        alignment: Alignment.center,
+        decoration: BoxDecoration(
+            color: color, borderRadius: BorderRadius.circular(15.r)),
+        padding: EdgeInsets.symmetric(
+          vertical: 15.h,
+        ),
+        child: Text(
+          "Get Started",
+          style: TextStyle(color: textColor),
+        ),
+      ),
     );
   }
 }
