+++ lib/widgets/content_sigin.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:online_course/utils/app_colors.dart';
+import 'package:online_course/widgets/button_push.dart';
+import 'package:online_course/widgets/inputs/custom_text_field.dart';
 import 'package:online_course/widgets/inputs/password_input.dart';
 import 'package:online_course/widgets/texts.dart';
 
@@ -54,37 +56,16 @@ class _ContentSiginState extends State<ContentSigin> {
             children: [
               Sub_text(text: "Recovery Password", size: 12.sp),
             ],
-          )
+          ),
+          SizedBox(
+            height: 24.h,
+          ),
+          ButtonPush(
+            text: "Sign In",
+            // color: AppColors.white_Color,
+          ),
         ],
       ),
-    );
-  }
-}
-
-class CustomTextField extends StatelessWidget {
-  final String hintText;
-  final TextInputType keyboardType;
-
-  CustomTextField({
-    required this.hintText,
-    this.keyboardType = TextInputType.text,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      decoration: InputDecoration(
-        contentPadding: EdgeInsets.symmetric(horizontal: 14.w, vertical: 16.h),
-        hintText: hintText,
-        hintStyle:
-            TextStyle(color: AppColors.input_text_Color, fontSize: 14.sp),
-        border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(15),
-            borderSide: BorderSide.none),
-        filled: true,
-        fillColor: Color(0xffF7F7F9),
-      ),
-      keyboardType: keyboardType,
     );
   }
 }
